UC-12 refactor: change page display folder
